Unify shared behaviour of the two response handlers.
